Add some test HTML tags to the database messages
An idea I had: on the backend store messages in a markup language. No
schema, but pick the format/tags so it renders nice on clients by
default.

When messages go, (e.g from server to client for display, server to
client for editing, or client to server for updating), the message
gets run through the appropriate administrator-definable 'message
processing pipeline'.

A pipeline is just a bunch of stream functions (each one operates on
one stream of characters and pushes out to another stream), just like
piping text between programs in unix. This mechanism seems flexible
and composible enough for most purposes.

For example, if users wanted a custom plugin where typing
`\time-of-day` did something, they'd make a pipeline that notices this
text and replaces it with something else for storage (perhaps
"<time-of-day />"). On the server to client pipeline they'd want to
replace <time-of-day /> with something like "12 oclock".

There would be different pipelines for different clients and contexts.
For example an IRC client would want most markdown stripped. And a
user wanting to edit a message would want to see the editing macro and
not the final result, etc.